adding ability for float

+++ integers/adder.go
@@ -1,5 +1,5 @@
 package integers
 
-func Add(a int, b int) int {
+func Add(a float32, b float32) float32 {
 	return a + b
 }
